This version fixes a few bugs:
* Can't view saved playlists if there is a file in the root directory.  
The parser has been improved to avoid this issue.

*  When clicking on the album cover for the current song in the top 
toolbar, it shows the album in the browser.  Previously, you got a 
default album icon instead of the cover art, it will now display the 
album art based on the tags of the first song in the list.

*  I don't think anyone else noticed this, but if you closed firefox 
while connected to a server, and then shut down that server, firefox 
would hang when starting the next time.  This is because it would 
continuously try to connect to the missing server.  It will now fail 
gracefully on any errors and not try again until you tell it to connect 
to the server again.